+++ src/Framework/tests/updaters/releaseupdater.test.ts
@@ -10,7 +10,7 @@ import { TaskDefinition } from "azure-devops-node-api/interfaces/TaskAgentInterf
 
 import { IGroupPermission, IReleasePermission, ITask, PermissionType } from "../../readers/iconfigurationreader";
 import { IConsoleLogger } from "../../common/iconsolelogger";
-import { IDebugLogger } from "../../common/idebuglogger";
+import { IDebugLogger } from "../../loggers/idebuglogger";
 import { IHelper } from "../../common/ihelper";
 import { IReleaseHelper } from "../../helpers/ireleasehelper";
 import { IReleaseUpdater } from "../../updaters/ireleaseupdater";
