+++ src/Framework/updaters/buildupdater.ts
@@ -6,7 +6,7 @@ import { IBuildHelper } from "../helpers/ibuildhelper";
 import { IBuildUpdater } from "./ibuildupdater";
 import { IBuildPermission } from "../readers/iconfigurationreader";
 import { IConsoleLogger } from "../common/iconsolelogger";
-import { IDebugLogger } from "../common/idebuglogger";
+import { IDebugLogger } from "../loggers/idebuglogger";
 import { IHelper } from "../common/ihelper";
 import { INamespace, ISecurityHelper, ISecurityIdentity } from "../helpers/isecurityhelper";
